requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
googlemaps==4.10.0
geopy==2.4.1
python-dotenv==1.0.0
Flask-CORS==4.0.0
